@@ -5,8 +5,7 @@ homepage: https://github.com/ponnamkarthik/drm_video
 issue_tracker: https://github.com/ponnamkarthik/drm_video/issues
 
 environment:
-  sdk: ">=2.7.0 <3.0.0"
-  flutter: ">=1.20.0"
+  sdk: ">=2.17.0-150.0.dev <3.0.00"
 
 dependencies:
   flutter:
